Added stuff and removed stuff and put stuff together
