+++ src/backend/controllers/MatchController.ts
@@ -1,24 +1,25 @@
 import { randomUUID } from "crypto";
-import { ClientMessage, createLobbyMessage, GameActionMessage, joinLobbyMessage, ServerMessage } from "../../interfaces/interfaces.js";
-import { GameLobby } from "../lobbies/GameLobby.js";
+import { ClientMessage, createLobbyMessage, GameActionMessage, joinLobbyMessage, ReadyMessage, ServerMessage } from "../../interfaces/interfaces.js";
 import { Player } from "../gamelogic/components/Player.js";
 import { MessageHandlers } from "../services/MessageHandlers.js";
 import { UserService } from "../services/UserService.js";
 import { WebSocket } from "ws";
 import { FastifyReply, FastifyRequest } from "fastify";
 import { MatchLobby } from "../lobbies/MatchLobby.js";
-// import { TournamentLobby } from "../lobbies/TournamentLobby.js";
+import { MatchService } from "../services/MatchService.js";
 
-export abstract class MatchController {
+export class MatchController {
     protected _lobbies: Map<string, MatchLobby>;
     protected _clients: Map<WebSocket, Player | null>; //
     protected _handlers: MessageHandlers;
     protected _userService: UserService;
     protected _invites: Map<string, { from: number, to: number, lobbyId: string, expires: Date }>
+    private _matchService: MatchService;
 
     constructor(userService: UserService, lobbies: Map<string, MatchLobby>) {
         this._userService = userService;
         this._lobbies = lobbies;
+        this._matchService = new MatchService(userService);
         this._clients = new Map<WebSocket, Player | null>();
         this._handlers = new MessageHandlers(this.broadcast.bind(this));
         this._invites = new Map<string, { from: number, to: number, lobbyId: string, expires: Date }>();
@@ -100,12 +101,25 @@ export abstract class MatchController {
             case "declineInvite":
                 this.handleDeclineInvite(connection, data.userId, data.inviteId)
                 break;
+            case "ready":
+                this.handlePlayerReady(connection, player!, (data as ReadyMessage).ready)
+                break;
+            case "startGame":
+                this.handleStartGame(connection, player!);
+                break;
+            case "pauseGame":
+                this.handlePauseGame(connection, player!);
+                break;
+            case "resumeGame":
+                this.handleResumeGame(connection, player!)
+                break;
+            case "getLobbyList":
+                this.handleGetLobbyList(connection);
+                break;
             default:
-                this.handleSpecificMessage(data, connection, player!);
+                throw Error("WTF DUDE!!!");
         }
     }
-
-    protected abstract handleSpecificMessage(data: ClientMessage, connection: WebSocket, player: Player | null): void
 
     private handleInvite(connection: WebSocket, fromUserId?: number, toUserId?: number) {
         if (!fromUserId || !toUserId) {
@@ -285,8 +299,6 @@ export abstract class MatchController {
         }
     }
 
-    protected abstract createLobby(lobbyId: string): GameLobby //| TournamentLobby
-
     protected handleJoinLobby(connection: WebSocket, userId?: number, lobbyId?: string) {
         if (!lobbyId) {
             this.sendMessage(connection, {
@@ -340,19 +352,143 @@ export abstract class MatchController {
         })
     }
 
-    public async getLobbies(request: FastifyRequest, reply: FastifyReply) {
-        const lobbies = [];
-
-        for (const [id, lobby] of this._lobbies.entries()) {
-            if (!lobby.isFull() && !lobby.isGameStarted()) {
-                lobbies.push({
-                    id: id,
-                    players: lobby.getPlayerCount(),
-                    maxPlayers: 2,
-                    creator: lobby.getCreatorId() || "Unknown"
-                })
-            }
+    private handlePlayerReady(connection: WebSocket, player: Player, isReady: boolean) {
+        if (!player || !player.lobbyId) {
+            this.sendMessage(connection, {
+                type: "error",
+                message: "not in a lobby"
+            })
+            return
         }
-        reply.code(200).send({ lobbies });
+
+        const lobby = this._lobbies.get(player.lobbyId) as MatchLobby
+        if (!lobby) {
+            this.sendMessage(connection, {
+                type: "error",
+                message: "Lobby not found"
+            })
+            return;
+        }
+        lobby.setPlayerReady(player.id, isReady)
+    }
+
+    private handleStartGame(connection: WebSocket, player: Player) {
+        if (!player || !player.lobbyId) {
+            this.sendMessage(connection, {
+                type: "error",
+                message: "not in lobby"
+            })
+            return;
+        }
+
+        const lobby = this._lobbies.get(player.lobbyId) as MatchLobby;
+        if (!lobby) {
+            this.sendMessage(connection, {
+                type: "error",
+                message: "lobby not found"
+            })
+            return;
+        }
+
+        if (lobby.getCreatorId() !== player.userId) {
+            this.sendMessage(connection, {
+                type: "error",
+                message: "Only creator can start game"
+            })
+            return;
+        }
+
+        if (lobby.getPlayerCount() < 2) {
+            this.sendMessage(connection, {
+                type: "error",
+                message: "Need at least 2 players"
+            })
+            return;
+        }
+
+        if (!lobby.checkAllPlayersReady()) {
+            this.sendMessage(connection, {
+                type: "error",
+                message: "All player must be ready"
+            })
+            return;
+        }
+
+        lobby.startGame();
+    }
+
+    private handlePauseGame(connection: WebSocket, player: Player) {
+        if (!player || !player.lobbyId) { return; }
+
+        const lobby = this._lobbies.get(player.lobbyId) as MatchLobby
+        if (lobby && lobby.getCreatorId() === player.userId) {
+            lobby.pauseGame();
+        } else {
+            this.sendMessage(connection, {
+                type: "error",
+                message: "Only lobby creator can pause game"
+            })
+        }
+    }
+
+    private handleResumeGame(connection: WebSocket, player: Player) {
+        if (!player || !player.lobbyId) { return; }
+
+        const lobby = this._lobbies.get(player.lobbyId) as MatchLobby
+        if (lobby && lobby.getCreatorId() === player.userId) {
+            lobby.resumeGame();
+        } else {
+            this.sendMessage(connection, {
+                type: "error",
+                message: "Only lobby creator can resume game"
+            })
+        }
+    }
+
+    private async handleGetLobbyList(connection: WebSocket) {
+        const openLobbies = await this._matchService.getOpenLobbies();
+
+        const dbLobbies = openLobbies.map( Lobby => {
+            const activeLobby = Array.from(this._lobbies.values()).find(l =>
+                l.getGameId() === Lobby.id
+            );
+
+            if (activeLobby)
+            {
+                return activeLobby.getLobbyInfo();
+            }
+
+            return {
+                id: Lobby.id.toString(),
+                name: Lobby.lobbyName || `Lobby ${Lobby.id}`,
+                creatorId: Lobby.player1.id,
+                maxPlayers: Lobby.maxPlayers || 2,
+                currentPlayers: Lobby.lobbyParticipants?.length || 1,
+                isPublic: !Lobby.hasPassword,
+                hasPassword: Lobby.hasPassword || false,
+                createdAt: Lobby.createdAt,
+                lobbyType: 'game' as const,
+                isStarted: false
+            }
+        })
+
+        const allLobbies = [...dbLobbies]
+        for (const [id, lobby] of this._lobbies.entries())
+        {
+            allLobbies.push(lobby.getLobbyInfo());
+        }
+
+        this.sendMessage(connection, {
+            type: "lobbyList",
+            lobbies: allLobbies
+        })
+    }
+
+    protected createLobby(lobbyId: string): MatchLobby {
+        return new MatchLobby(
+            lobbyId,
+            this.broadcast.bind(this),
+            this._matchService
+        )
     }
 }
