--- src/backend/models/GameModel.ts
@@ -1,4 +1,0 @@
-// import { ChildEntity, Column, CreateDateColumn, Entity, JoinColumn, JoinTable, ManyToMany, ManyToOne, PrimaryGeneratedColumn } from "typeorm";
-// import { MatchModel } from "./common.js"
-
-// const { UserModel } = require("./common.js") as typeof import("./common.js");
