+++ src/backend/models/MatchModel.ts
@@ -94,10 +94,6 @@ export class MatchModel {
 
     @Column({ type: 'datetime', nullable: true })
     endedAt?: Date;
-}
-
-@ChildEntity("games")
-export class GameModel extends MatchModel {
 
     @Column({ default: false })
     isLobbyOpen!: boolean;
@@ -127,3 +123,110 @@ export class GameModel extends MatchModel {
     @Column("simple-array", { nullable: true })
     invitedUserIds?: number[];
 }
+
+// @ChildEntity("games")
+// export class GameModel extends MatchModel {
+
+//     @Column({ default: false })
+//     isLobbyOpen!: boolean;
+
+//     @ManyToMany(() => UserModel)
+//     @JoinTable()
+//     lobbyParticipants!: typeof UserModel.prototype[];
+
+//     @Column({ nullable: true })
+//     gameAdminId?: number;
+
+//     @Column({ default: 2 })
+//     maxPlayers!: number;
+
+//     @Column({ default: '' })
+//     lobbyName!: string;
+
+//     @Column({ default: false })
+//     hasPassword!: boolean;
+
+//     @Column({ nullable: true })
+//     passwordHash?: string;
+
+//     @Column('simple-json', { nullable: true })
+//     readyStatusMap?: Record<number, boolean>;
+
+//     @Column("simple-array", { nullable: true })
+//     invitedUserIds?: number[];
+// }
+
+
+
+// @ChildEntity("tournaments")
+// export class TournamentModel extends MatchModel {
+//   @Column({ default: 8 })
+//   maxParticipants!: number;
+
+//   @ManyToMany(() => UserModel)
+//   @JoinTable()
+//   participants!: UserModel[];
+
+//   @ManyToMany(() => GameModel)
+//   @JoinTable()
+//   matches!: GameModel[];
+
+//   @Column({ default: 'registration' })
+//   tournamentPhase!: 'registration' | 'in_progress' | 'completed';
+// }
+
+
+// @Entity('tournament_matches')
+// export class TournamentMatchModel {
+//     @PrimaryGeneratedColumn()
+//     id!: number;
+
+//     @Column()
+//     tournamentId!: number;
+
+//     @ManyToOne(() => TournamentModel, tournament => tournament.matches)
+//     @JoinColumn({ name: 'tournamentId' })
+//     tournament!: TournamentModel;
+
+//     @ManyToOne(() => UserModel)
+//     @JoinColumn({ name: 'player1Id' })
+//     player1!: UserModel;
+
+//     @Column()
+//     player1Id!: number;
+
+//     @ManyToOne(() => UserModel)
+//     @JoinColumn({ name: 'player2Id' })
+//     player2!: UserModel;
+
+//     @Column()
+//     player2Id!: number;
+
+//     @Column({ default: 0 })
+//     player1Score!: number;
+
+//     @Column({ default: 0 })
+//     player2Score!: number;
+
+//     @ManyToOne(() => UserModel, { nullable: true })
+//     @JoinColumn({ name: 'winnerId' })
+//     winner!: UserModel;
+
+//     @Column({ nullable: true })
+//     winnerId!: number;
+
+//     @Column({ default: 'pending' })
+//     status!: 'pending' | 'ongoing' | 'paused' | 'completed' | 'cancelled';
+
+//     @Column()
+//     matchNumber!: number;
+
+//     @CreateDateColumn()
+//     createdAt!: Date;
+
+//     @Column({ nullable: true, type: 'timestamp' })
+//     startedAt!: Date;
+
+//     @Column({ nullable: true, type: 'timestamp' })
+//     endedAt!: Date;
+// }
