--- src/backend/models/UserModel.ts
@@ -1,1 +1,0 @@
-// import { Column, Entity, JoinColumn, JoinTable, ManyToMany, ManyToOne, OneToMany, PrimaryGeneratedColumn } from "typeorm";
