--- src/backend/models/common.ts
@@ -1,4 +1,0 @@
-// export {UserModel} from './UserModel.js'
-// export {MatchModel} from './MatchModel.js'
-// export {GameModel} from './GameModel.js'
-// // export {TournamentMatchModel} from './TournamentModel.js'
